third_party/WebKit/LayoutTests/inspector/sources/inspect-function.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>
function test()
{
    var revealed = 0;
    InspectorTest.addSniffer(Sources.SourcesView.prototype, "showSourceLocation", didReveal, true);
    InspectorTest.evaluateInConsole("function foo() { }; inspect(foo.bind());inspect(foo);");

    function didReveal(uiSourceCode, lineNumber, columnNumber)
    {
        InspectorTest.addResult("Function was revealed:");
        InspectorTest.addResult("lineNumber: " + lineNumber);
        InspectorTest.addResult("columnNumber: " + columnNumber);
        if (revealed == 0)
            ++revealed;
        else
            InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
    <p>Tests that inspect object action works for function and resolve bound function location.</p>
</body>
</html>
