third_party/WebKit/LayoutTests/inspector/sources/inspect-function-expected.txt

Tests that inspect object action works for function and resolve bound function location.

Function was revealed:
lineNumber: 1
columnNumber: 12
Function was revealed:
lineNumber: 1
columnNumber: 12
